backend/requirements.txt
'''fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.196
scipy==1.11.4
matplotlib==3.7.2'''


fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic[email]==2.5.0
bcrypt==4.0.1
PyJWT==2.8.0
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.196
scipy==1.11.4
matplotlib==3.7.2
